extra http error processing from errorModal to custom hook
