+++ src/components/UI/Modal/Modal.js
@@ -25,4 +25,6 @@ Modal.propTypes = {
     remove: PropTypes.func
 }
 
-export default React.memo(Modal, (prevProps, nextProps) => nextProps.show === prevProps.show && nextProps.children === prevProps.children);
+export default React.memo(Modal, (prevProps, nextProps) =>
+    nextProps.show === prevProps.show &&
+    nextProps.children === prevProps.children);
